m7-common: consolidate GSM device proprietary files under m7-gsm

+++ extract-files.sh
@@ -21,18 +21,20 @@ fi
 BASE=../../../vendor/$VENDOR/$DEVICE/proprietary
 rm -rf $BASE/*
 
-for FILE in `egrep -v '(^#|^$)' ../$DEVICE/device-proprietary-files.txt`; do
-  echo "Extracting /system/$FILE ..."
-  DIR=`dirname $FILE`
-  if [ ! -d $BASE/$DIR ]; then
-    mkdir -p $BASE/$DIR
-  fi
-  if [ "$SRC" = "adb" ]; then
-    adb pull /system/$FILE $BASE/$FILE
-  else
-    cp $SRC/system/$FILE $BASE/$FILE
-  fi
-done
+if [ -f ../$DEVICE/device-proprietary-files.txt ]; then
+  for FILE in `egrep -v '(^#|^$)' ../$DEVICE/device-proprietary-files.txt`; do
+    echo "Extracting /system/$FILE ..."
+    DIR=`dirname $FILE`
+    if [ ! -d $BASE/$DIR ]; then
+      mkdir -p $BASE/$DIR
+    fi
+    if [ "$SRC" = "adb" ]; then
+      adb pull /system/$FILE $BASE/$FILE
+    else
+      cp $SRC/system/$FILE $BASE/$FILE
+    fi
+  done
+fi
 
 for FILE in `egrep -v '(^#|^$)' ../m7-common/proprietary-files.txt`; do
   echo "Extracting /system/$FILE ..."
